pyproject.toml
[tool.poetry]
name = "mnf"
version = "0.1.0"
description = "A flexible python application to forward Mac OS Notifications to Other Platforms"
authors = ["Sean Han <hanxiaobow@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.27.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
